requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-google-genai
google-generativeai
python-multipart
python-dotenv